backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2
python-dotenv==1.0.1
openai==1.40.0
IMAPClient==2.3.1
beautifulsoup4==4.12.3
html2text==2020.1.16
APScheduler==3.10.4
requests==2.32.3
